requirements_all.txt
-r ./requirements.txt
docker==4.2.0
Flask-APScheduler==1.11.0
locust~=1.4.3
pytest~=5.2.2
PySnooper~=0.4.0
Flask-RESTful==0.3.8
requests-toolbelt~=0.9.1
cchardet~=2.1.5
fleep~=1.0.1
ffmpeg-python~=0.2.0
natsort~=7.0.1

pydub~=0.24.1
aiohttp==3.8.3
zhon~=1.1.5

chardet~=3.0.4
six~=1.15.0
websocket-client~=0.57.0

beautifulsoup4~=4.9.3
pycryptodome~=3.10.1

# rasa
boto3==1.17.49
Elasticsearch~=7.12.1
minio==7.0.3

celery~=5.2.1
redis~=4.4.0
xlrd==1.2.0

# pymilvus
pymilvus~=2.0.0